Update hyperspawn.sp
Allows for equipping players after they spawn/teleport. The array is cleared and re-populated every time so it can allow for additional info_player_equip's to be created during the map.
Supposedly, we could check if the env_global IsVehicleMap is set to 1 and only additionally equip when it is. But there should be no harm in additionally equipping.